Initial Playwright testing.
In order to properly run Playwright tests in this monorepo, I had to
set "type": "module" in the root `packge.json`.  Without that I was getting
an error saying that the ".ts" was an unknown file extension.
This answer tipped-me off on this solution:
https://github.com/microsoft/playwright/issues/18282#issuecomment-1315834051

For these initial Playwright tests, I created a temporary dummy page
and added a ".e2e.ts" file that is associated with that file.